Moved Block to a separate module, +clippy
